pyproject.toml
[project]
name = "visual-odometry"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    # uv add <package> will automatically add here
    "pre-commit>=4.4.0",
    "rerun-sdk>=0.27.2",
    "scipy>=1.16.3",
    "tyro>=0.9.35",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit",
]



# Mypy - Type Checking
[tool.mypy]
python_version = "3.11"
mypy_path = "src"
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
warn_redundant_casts = true
show_error_codes = true
# Disabled strict options for easier collaboration
# disallow_untyped_defs = false  # Allow functions without type hints
# no_implicit_optional = false   # Don't require Optional[] everywhere

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest - Testing
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=visual_odometry",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage.py
[tool.coverage.run]
branch = true
source = ["visual_odometry"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

# Ruff - Fast Python Linter (Alternative to Flake8)
# Uncomment if you want to use ruff instead of flake8
[tool.ruff]
line-length = 88
target-version = "py311"
[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "F401",  # unused imports
    "T201",  # print statements found
    "D100",  # missing docstring in public module
    "N803",  # argument name should be lowercase
    "PIE790",  # pass statement found
    "D212",  # multi-line docstring summary should start at the first line
    "D203",  # multi-line docstring closing quotes should be on a separate line
    "PLR0913",  # too many arguments
    "TD002",  # missing author in docstring
    "TD003",  # missing issue link in docstring
    "FIX002",  # fixme found
    "ERA001", # found commented out code
    "E501",  # line too long
    "D417",  # missing argument descriptions in docstring
    "PLR2004",  # magic value found
    "COM812",  # missing trailing comma in a collection
    "N806",  # variable in function should be lowercase
    "D104",  # missing docstring in public package
    "ARG001",  # unused function argument
    "RUF059", # unpacked variable is never used
]
fixable = ["ALL"]
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.format]
exclude = ["Q000", "D212"]
skip-magic-trailing-comma = false
